adding modal backgournd opacity and separating components into individual files (except terms)
C-Co-Authored-by: develiasdaniel eliasdanielsanchezlopez@gmail.com
